+++ src/main/java/com/schoolsaas/dto/quiz/QuizResultDto.java
@@ -2,6 +2,7 @@ package com.schoolsaas.dto.quiz;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -10,9 +11,9 @@ public class QuizResultDto {
     private UUID submissionId;
     private UUID quizId;
     private String quizTitle;
-    private Double score;
-    private Double totalMarks;
-    private Double percentage;
+    private BigDecimal score;
+    private BigDecimal totalMarks;
+    private BigDecimal percentage;
     private String gradeLetter;
     private String status;
     private List<QuizAnswerDto> answers;
