+++ src/main/java/com/schoolsaas/dto/reportcard/ReportCardDto.java
@@ -2,6 +2,7 @@ package com.schoolsaas.dto.reportcard;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -17,8 +18,8 @@ public class ReportCardDto {
     private Integer attendancePresent;
     private Integer attendanceAbsent;
     private Integer attendanceLate;
-    private Double totalScore;
-    private Double averageScore;
+    private BigDecimal totalScore;
+    private BigDecimal averageScore;
     private String overallGrade;
     private Integer classPosition;
     private Integer classSize;
